+++ package.json
@@ -36,8 +36,6 @@
     "react-native-reanimated": "^2.14.0",
     "react-native-safe-area-context": "^4.4.1",
     "react-native-screens": "^3.18.2",
-    "react-native-swipe-gestures": "^1.0.5",
-    "react-native-swipe-list-view": "^3.2.9",
     "react-native-swipeable-list": "^0.1.2",
     "react-native-toast-message": "^2.1.5",
     "react-native-tab-view": "^3.3.4",
